Add support for parsing complex function docstings. (#175)
* Add support for multiline endpoint summary and description strings.

Instead of the first new line character, use the first empty line as
the separator between the summary and description section in a
function docstring. This enables breaking up long summary and description
sections into multiple lines if horizontal space is limited.

* Truncate function doscrings during parsing, after a form feed character.

Function docstrings, beyond the basic summary and description, frequently
contain additional details related to parameter and return value
documentation, that is not very useful when it shows up on an API
documentation page.
By introducing a special character, that rarely if ever shows up in a
function docstring, we can mark a section of the docstring to
be ignored by the parser.
In our case, the special character that can be used for this purpose is
chosen to be the \f (0xC), form feed character. (This convention has
been copied from FastAPI
https://fastapi.tiangolo.com/advanced/path-operation-advanced-configuration/#advanced-description-from-docstring.)